Move base actions to models dir
